+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+
+## Ver 4.0.6
+
+Fix for whitespace from being stripped during normal text insertion
+
 ## Ver 4.0.5
 
 New:
@@ -7,6 +12,7 @@ New:
 - Media seeking to cursor position (+ documentation)
 - Insert text from file
 - Normal Paste: Insert text through copying from another application 
+
 Change:
 
 - Index text now uses a non-breaking space